Use libphodav-2 (breaks webdav server temporarily)
This change breaks webdav server, since libphodav-2 no longer
set up a TCP service running in a thread. It's up to the client
to decide how best to accept and handle new connections.

This commits remove all the hacks related to proxying the incoming
connections to a TCP socket, and protected with a magic sequence.

The following commit will use GIOStream pipes to handle each client
connections.