channel: use exisiting reconnection code when switching
Switching for migration reason is similar to a reconnection during
initial connection.

A notable difference is that new code path doesn't schedule a
delayed_unref callback. This is fine since the channel is still running
and delayed_unref is mainly used for signaling disconnections and none
should be emitted when switching.