session: reset connection_id on switch-host
The server expects no connection_id during link time, in order to
realize a full reconnection.

The alternative would be to assume that the destination server has the
exact same channels and attempt to recreate and to reconnect them one by
one. However, if the destination is slightly different (say,
configuration or order of channels differs), this will likely fail. It's
best to start with a new session without prior knowledge.