audio: spice-pulse implement async volume-info
In case of volume-sync between client and guest, we request volume-info
from the availables streams and if the stream is not available we rely
on ext-stream-restore.

By using ext-stream-restore we can get the last stream data of the
application that is stored by PulseAudio.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=1012868