channel: HUP is not an error condition
On local UNIX sockets, calling close() in the server side will result in
a HUP condition on client side. Since this is not an error, but a normal
termination, let's ignore it.

A clean shutdown would involve a new message to tell the client to first
close its end, in order to avoid the HUP. That way the client could
distinguish normal termination from unexpected ones. That's a possible
future minor enhancement (it seems it would only work with UNIX socket
though)

G_IO_IN seemed to be a leftover, it is left here but shouldn't be
required.

https://bugs.freedesktop.org/show_bug.cgi?id=90016