session: disable default socket proxy
GSocketClient uses the system proxy by default, and it may end up using
the system HTTP proxy with bad results if CONNECT is not
supported. spice-gtk uses it's own SPICE_PROXY instead, and doesn't rely
on GSocketClient default proxy but GProxyAddress instead. Disabling the
default proxy solve a wrong proxy from being used.

It may be worth to revisit this change if GSocketClient can be told to
ignore proxies that are not eligible for Spice connections. (HTTP could
be though, in which case it would be a user configuration issue)

Fixes:
https://bugzilla.redhat.com/show_bug.cgi?id=1040679