This adds reference counting to cached images.
Windows guests with multi-monitors will often send down the same image
via different channels.  If these instances are not counted,  one channel
can delete an image before all the rest of the channels are finished with
it. In this case, the client will hang.

This can happen, for instance, when the Windows guest is updating the
monitor's wallpaper. There is a driver instance for each monitor, so each
one is sending down its own copy of the same image with the same id.  As
one channel may be busier than another, the order that the client executes
commands may not be the same order as the server issued them.  Here's what
can happen:

  On the server side:
    Monitor 1 sends image A
    Monitor 1 sends multiple rendering commands for image A
    Monitor 1 removes image A
    Monitor 2 sends image A
    Monitor 2 sends rendering commands for image A
  On the client side:
    Monitor 1 adds image A to the cache
    Monitor 1 starts rendering with image A
    Monitor 2 adds image A to the cache - has same id, so image is replaced
    Monitor 1 removes image A - image A is completely removed from cache
    Monitor 2 render command hangs waiting for image A

Addresses bug: https://bugzilla.redhat.com/show_bug.cgi?id=1194354