Call spice_marshaller_flush on message write
spice_marshaller_flush must be called before writing a message to
calculate pointer offsets.

This is only an issue when ptr submarshallers are used, which currently
is only used by Tunnel::service_add. Since this is disabled by default,
this buug is not going to be hit by the current codebase.