Add the gtk libraries to the gtk-scandoc link line.
This prevents a compile error on Debian Jessie, when building from git.

This is fairly subtle, and Debian specific.  It only happens when you use
autoreconf to generate a new libtool script.  Debian patches that script
to require an explicit setting to link with all dependent libraries.

It should be harmless on other distros, and it does save us Debian guys some
hassle.