build-sys: Fix spice-protocol requirements in .pc file
Commit d0ae58a8 added spice-protocol >= 0.12.10 to the Requires.private
field in spice-client-glib.pc. However it's already present in the
Requires field without a version check. Once spice-gtk is built and
installed, building a program against spice-gtk needs spice-protocol,
but no specific version of it, so we can drop the duplicate
Requires.private requirement, and keep the unversioned one in Requires.