channel-usbredir: drop isoc packets on low bandwidth
When channel wants to send much more data then the wire can handle, the
queue grows fast. This patch does not limit the queue growth but
introduces an internal API to check if queue size is too big.

This internal API is used in usbredir_buffered_output_size_callback
which is called before any isoc pacaket is queued in usbredir. The
usbredir implements the logic to:
 - only drop isoc packets
 - while dropping packtes does still give us video frames from and above
   10fps streams

An easy way to test locally is sharing and webcam and with tc:
    tc qdisc add dev lo root netem delay 100ms
    tc qdisc change dev lo root netem delay 1000ms
    tc qdisc del dev lo root netem

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1264156