Use #ifdef {HAVE,USE}_FOO instead of #if {HAVE,USE}_FOO
While "#if USE_FOO" checks for the value of the variable USE_FOO,
"#ifdef" checks whether USE_FOO is defined or not.

It means, if we had something like: #define USE_FOO 0, we would have:
 #if USE_FOO
 /* Any code in here would NOT be compiled */
 #endif

 #ifdef USE_FOO
 /* Any code in here would be compiled */
 #endif

No problem was faced on spice-gtk till now because either USE_FOO is not
defined or defined as 1, but let's try to have it standardized.

Acked-by: Eduardo Lima (Etrunko) <etrunko@redhat.com>