main: don't update display timer for unchanged config
With virgl, set_monitor_ready() may be called each time the scanout is
updated to set the monitor area. This will call
spice_main_update_display(), and keep the timer postponed even if the
monitor configuration didn't change. Treat unchanged configuration as a
no-op and keep configuration timer unchanged. This fixes monitor
autoconfig with virgl (when the display is regularly updated).

It also solves/avoids problems with a guest running on wayland when the
"resize-guest" property is TRUE.

Fixes:
https://bugzilla.redhat.com/show_bug.cgi?id=1266484

Signed-off-by: Marc-André Lureau <marcandre.lureau@gmail.com>
Acked-by: Jonathon Jongsma <jjongsma@redhat.com>