usb device widget: don't try to disconnect on failed connect
When you try to redirect a usb device to the guest and it fails, we
uncheck the checkbox for this device. This causes the 'clicked' signal
to be emitted, which causes us to try to disconnect the device (which is
not currently connected, since the connect operation failed).

When we try to disconnect an unconnected device, the device manager
leaks memory and gets left in an inconsistent state because we allocate
the task data, call _set_redirecting(self, TRUE) and then return early
with the following warning printed to the terminal:

(lt-spicy:4638): GSpice-CRITICAL **: spice_usbredir_channel_disconnect_device_async: assertion 'channel != NULL' failed

To avoid this, disable signal handlers for the checkbox 'clicked'
signal while we're changing the checkbox state in response to a
connection error. In addition, add an additional check to
spice_usb_device_manager_disconnect_device_async() to ensure that the
passed device is actually connected.