Spice-widget: Allow smaller widget with scaling enabled
When resize-guest-to-match-window-size is disabled, the size request
(minimum size) of the spice widget is set to the current guest
resolution. The client window cannot be made smaller, even with
scale-display enabled.

This patch sets a size request of 640x480 when either
resize-guest-to-match-window-size OR scale-display are enabled, making
it posible to shrink the window and scale down the display.

Acked-by: Pavel Grunt <pgrunt@redhat.com>