doc,gir: Don't skip spice_get_option_group()
There is no reason to not include the spice_get_option_group
method in the generated bindings, as it can be useful.

It was removed in the commit 2db9b8dd037e22d2b04e8e2aeecfd685524b7fef,
for the reason of causing warnings, but no warnings show up now.

Acked-by: Fabiano Fidêncio <fidencio@redhat.com>