egl: fix delayed widget realize
When the display is not yet realized, spice_display_widget_gl_scanout()
will fail because the egl context is not ready. The display is never
marked ready because the egl.image (and egl.scanout) is not set, and
some clients, such as virt-viewer will not realize the widget until the
display is ready.

Deal with gl scanout updates when the widget is not yet realized, and
mark the display as ready when egl is enabled (when last display draw
signal is from gl).

Signed-off-by: Marc-André Lureau <marcandre.lureau@gmail.com>
Acked-by: Pavel Grunt <pgrunt@redhat.com>