main: channel-main to increase file-transfer reference
This is a minor fix in the logic as in both situations (with or
without the patch) the reference count for the SpiceFileTransferTask
object is the same.

The change is interesting as SpiceFileTransferTask is created but on
g_file_read_async() it increases its reference count while
c->file_xfer_tasks keeps the original one.

It should be the other way around.
Acked-by: Marc-André Lureau <marcandre.lureau@gmail.com>