file-xfer: introduce functions to read file async
Introduced functions (private):
* void   spice_file_transfer_task_read_async()
* gssize spice_file_transfer_task_read_finish()

For a better abstraction of how to read from SpiceFileTransferTask and
handle its data, following the design of other objects like GFile and
GInputStream.

Due to the logic changes involved, some functions were created or
renamed to better address or match its place and purpose:

* spice_file_transfer_task_read_stream_cb
  Callback for the actual read from GInpustStream; This is handling
  the SpiceFileTransferTask bits only;

* file_xfer_read_cb -> file_xfer_read_async_cb
  Renamed to match _read_async() function; This is handling the data
  from reading the file by flushing it to the agent.

As the _read_async() uses GTask, the error handling is done on the
channel-main's callback, after _read_finish() is called.

This change is related to split SpiceFileTransferTask from
channel-main.

Acked-by: Jonathon Jongsma <jjongsma@redhat.com>