build: Fix _DEPENDENCIES use
We want to trigger rebuild of libspice-client-gtk-3.0.la or
libspice-client-glib-2.0.la whenever the corresponding symbol file
changes.
However _DEPENDENCIES is not the right way of handling that as it will
disable automatic automake dependency generation.
This was not causing issues mainly because _DEPENDENCIES was mispelt as
_DEPEDENCIES.

Quoting automake manual:
https://www.gnu.org/software/automake/manual/automake.html#index-EXTRA_005fmaude_005fDEPENDENCIES-1
« The EXTRA_*_DEPENDENCIES variable may be useful for cases where you
merely want to augment the automake-generated _DEPENDENCIES variable
rather than replacing it. »

So this commit switches to use EXTRA_*_DEPENDENCIES rather than
*_DEPENDENCIES.