main: do not check for xfer-task errors
The errors related to SpiceFileTransferTask can be present:
* before the flush happens, on spice_file_transfer_task_read_async()
* during the async flush:
  - cancel from user which is handled by file_xfer_flush_async
  - cancel/error from agent, handled on main_agent_handle_xfer_status()

file_xfer_flush_finish() should not check internal errors of
SpiceFileTransferTask and only be worried about errors regarding the
flush itself.

Note that with or without this patch, in case of errors from agent, we
don't stop the flushing; it is only cancelled from user cancellation.

Acked-by: Jonathon Jongsma <jjongsma@redhat.com>