main: let channel-main handle file-xfer messages
By separating SpiceFileTransferTask from channel-main, we could choose
where to put the handler for messages. With the approach based on
spice_file_transfer_task_read_async(), we cannot have it under
spice-file-transfer-task.c due the need of callback and userdata on
_read_async.

It is much easier to keep this in channel-main and do not move any
VDAgent.

This patch reverts 349a52ca2d6af4b31a4f51c38a3292c04460953c changes
but it does rename:
- function file_xfer_handle_status -> main_agent_handle_xfer_status
- variables task to xfer_task

Acked-by: Jonathon Jongsma <jjongsma@redhat.com>