channel-main: improve code to use FileTransferOperation
Since commit b26818d0e00666 we are grouping all transferred files per
operation (drag and drop); That leaded to some design flaws in the
code. The changes suggested in this patch are (in execution order):

* On spice_file_transfer_task_read_async()
- Passing FileTransferOperation data so in the callback and in further
  function calls, we don't need to look for this;

Note that FileTransferOperation is not freed while any of its
SpiceFileTransferTask exists and by design they are never finished
while on pending state, so file_xfer_read_async_cb() should have a
valid FileTransferOperation pointer.

* On file_xfer_read_async_cb()
- Removed unnecessary variables;

* On file_xfer_flush_async()
- Using SpiceFileTransferTask as parameter which can retrieve the
  SpiceMainChannel and the GCancellable;
- Using SpiceFileTransferTask as source_object for GTask with
  FileTransferOperation as user_data which is helpful for its
  callback;

* On file_xfer_flush_finish() and file_xfer_data_flushed_cb()
- Using SpiceFileTransferTask as parameter and source_object check for
  GTask

Acked-by: Jonathon Jongsma <jjongsma@redhat.com>